fix: reader correctness - lookup wedge, hotbar name cache, unmapped-field sentinels, total gauge/player reads
- ActionLookup/StatusEffectLookup/ZoneLookup: atomic check-and-set on the
  in-flight flag plus try/finally. One Lumina exception (likely in the window
  between a game patch and the matching Lumina release) previously left
  _loading stuck true, so every name lookup returned '???' until restart.
- Hotbar cache no longer overwrites the live action name on a cache hit -
  after a job/page swap every bound slot kept its pre-swap name forever.
  The cache now stores only values derived from the keybind string itself.
- ActorItem.Status was aliased to StatusManager's base, surfacing the low
  byte of an Owner pointer as the actor state. Unmapped fields now use a -1
  sentinel (guards skip them; TryTo* returns default for negatives) instead
  of 0, which silently read the vtable byte.
- ActorItem.SourceSize derives from BattleChara (0x3810), not Character
  (0x2370): CastInfo/StatusManager offsets exceed sizeof(Character) and only
  worked via ArrayPool bucket rounding. Same 16K bucket, so no read growth.
- JobResourceResolver/PlayerInfoResolver: every read is now total (guarded
  ReadByte + never-throwing SharlayanBitConverter with exact bounds checks) -
  a single bad offset zeroes one field instead of aborting the remaining
  gauges/fields via the shared catch.
- Reader.GameState BGM scenes: one bulk read of the Scene vector (was up to
  ~84 syscalls/poll) parsed into fresh BgmSceneInfo instances per poll - the
  previous shared cache mutated results a consumer might still hold.
- ResolvePointerPath ASM hop now aborts on a failed displacement read via
  TryGetInt32 (closes the asymmetry flagged in c52f19a's review).
- BardResources.ActiveSong offset widened byte->int (silent-wrap hazard).
- Mapper tests updated to pin the new contract (BattleChara size, -1
  sentinels); integrity test uses the file's alias convention.

+++ Sharlayan/Models/Structures/JobResources.Bard.cs
@@ -11,7 +11,9 @@
 namespace Sharlayan.Models.Structures {
     public partial class JobResources {
         public sealed class BardResources {
-            public byte ActiveSong { get; set; }
+            // F88: this is a byte OFFSET into the gauge struct like every sibling —
+            // int, so a future FCS layout push past 255 can't silently truncate.
+            public int ActiveSong { get; set; }
             public int Repertoire { get; set; }
             public int SoulVoice { get; set; }
             public int RadiantFinaleCoda { get; set; }
